requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mlflow==2.8.1
xgboost==2.0.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.1
pyspark==3.5.0
kafka-python==2.0.2
pydantic==2.5.0
optuna==3.4.0
shap==0.44.1
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pyyaml==6.0.1
requests==2.31.0
databricks-sdk==0.12.0
tenacity==8.2.3
torch==2.1.1
polars==0.19.19
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
jupyter==1.0.0